Suppression de warnings à la compilation sous Linux(gcc 8.3) et bonne lecture de l'OCP

+++ libft/ft_swap_uint16.c
@@ -19,5 +19,5 @@
 
 uint16_t					ft_swap_uint16(uint16_t nb)
 {
-	return (((nb & 0x00ff) >> 8) | ((nb & 0xff00) << 8));
+	return (((nb & 0x00ff) >> 8) | (uint16_t)((nb & 0xff00) << 8));
 }
